Login-in and register pages pull request #1
Added all necessary files (put them in folders) to login-in and register pages. In React files included html-layouts: login-in seemed to be ok, but register had some issues with styles (it was noted by green comment below). It was not included Burger code as I didn't understand properly how it works.